Changes cassandra dependency to 3.2.1 and cassandra driver to datastax driver 3.0.0
Changes query to enumerate cassandra keyspaces from system.schema.keyspaces to system_schema.keyspaces in CQLDataLoader
Changes isSystemKeyspace function in EmbeddedCassandraServerHelper to include new system_distributed and system_schema keyspaces
Changes Spring Unit tests which were asserting that the number of keyspaces available = 4, this is now 6 because of new keyspaces.
Changes cassandra-unit/src/main/resources/cu-cassandra.yaml to include hints directory, which is now required.
Changes DataLoaderTest and MockDataSetHelper, which were supplying min/max compaction thresholds to the LeveledCompactionStrategy, this are only compatible with the SizeTieredCompactionStrategy and are now being set on a different ColumnFamily wthin the same test.